requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
openai-whisper==20231117
python-multipart==0.0.9
pytube==15.0.0
google-generativeai==0.5.2